refactor: JPA Auditing을 사용하여 생성일 자동화 처리
