+++ src/main/java/weavers/siltarae/SiltaraeBeApplication.java
@@ -2,7 +2,9 @@ package weavers.siltarae;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class SiltaraeBeApplication {
