+++ src/main/java/weavers/siltarae/mistake/domain/Mistake.java
@@ -4,11 +4,10 @@ import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
+import weavers.siltarae.global.BaseEntity;
 import weavers.siltarae.tag.domain.Tag;
 import weavers.siltarae.user.domain.User;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 import static lombok.AccessLevel.PROTECTED;
@@ -16,7 +15,7 @@ import static lombok.AccessLevel.PROTECTED;
 @Entity
 @Getter
 @NoArgsConstructor(access = PROTECTED)
-public class Mistake {
+public class Mistake extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -28,12 +27,6 @@ public class Mistake {
     @Column(nullable = false, length = 140)
     private String content;
 
-    @CreatedDate
-    @Column(updatable = false, nullable = false)
-    private LocalDateTime createdAt;
-
-    private LocalDateTime deletedAt;
-
     @ManyToMany(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     @JoinTable(name = "tag_mistake",
             joinColumns = @JoinColumn(name = "mistake_id"),
@@ -41,15 +34,11 @@ public class Mistake {
     private List<Tag> tags;
 
     @Builder
-    public Mistake(Long id, User user, String content, LocalDateTime createdAt, List<Tag> tags) {
+    public Mistake(Long id, User user, String content, List<Tag> tags) {
         this.id = id;
         this.user = user;
         this.content = content;
-        this.createdAt = createdAt;
         this.tags = tags;
     }
 
-    public void deleteMistake(Mistake mistake) {
-        mistake.deletedAt = LocalDateTime.now();
-    }
 }
